+++ src/bs_gstl/src/sugarbox_neighbour_lookup_distance.h
@@ -31,7 +31,7 @@ namespace hpgl
 		neighbour_lookup_distance_t(
 				const neighbourhood_param_t & nb_param
 				)
-				:	m_max_neighbours(nb_param.m_max_neighbours), 
+				:	m_max_neighbours((int)nb_param.m_max_neighbours), 
 					m_vectors(new std::vector<sugarbox_vector_t>())
 		{
 			calc_dist_field<sugarbox_location_t>(nb_param.m_radiuses, *m_vectors);
@@ -49,7 +49,7 @@ namespace hpgl
 			const sugarbox_vector_t * vec = &vectors[0];
 			sugarbox_location_t center = grid[node];
 			node_coord = coord_t(center[0], center[1], center[2]);
-			for (int idx = 0, end_idx = vectors.size(); idx < end_idx && count < m_max_neighbours; ++idx, ++vec)
+			for (int idx = 0, end_idx = (int) vectors.size(); idx < end_idx && count < m_max_neighbours; ++idx, ++vec)
 			{
 				sugarbox_location_t point = center + *vec; //vectors[idx];
 				int index = grid.get_index(point);
